Updated paths after moving ui.js to ui folder
